+++ lib/coverage.dart
@@ -1,6 +1,7 @@
 library coverage;
 
-export 'src/hitmap.dart';
-export 'src/formatter.dart';
-export 'src/resolver.dart';
 export 'src/collect.dart';
+export 'src/formatter.dart';
+export 'src/hitmap.dart';
+export 'src/resolver.dart';
+export 'src/run_and_collect.dart';