+++ test/collect_coverage_test.dart
@@ -9,14 +9,16 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:coverage/coverage.dart';
+import 'package:coverage/src/util.dart';
 import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
 
-final _sampleAppPath = p.join('test', 'test_files', 'test_app.dart');
+import 'test_util.dart';
+
 final _isolateLibPath = p.join('test', 'test_files', 'test_app_isolate.dart');
 final _collectAppPath = p.join('bin', 'collect_coverage.dart');
 
-final _sampleAppFileUri = p.toUri(p.absolute(_sampleAppPath)).toString();
+final _sampleAppFileUri = p.toUri(p.absolute(testAppPath)).toString();
 final _isolateLibFileUri = p.toUri(p.absolute(_isolateLibPath)).toString();
 
 const _timeout = const Duration(seconds: 5);
@@ -71,15 +73,19 @@ void main() {
   test('parseCoverage', () async {
     var tempDir = await Directory.systemTemp.createTemp('coverage.test.');
 
-    var outputFile = new File(p.join(tempDir.path, 'coverage.json'));
+    try {
+      var outputFile = new File(p.join(tempDir.path, 'coverage.json'));
 
-    var coverageResults = await _getCoverageResult();
-    await outputFile.writeAsString(coverageResults, flush: true);
+      var coverageResults = await _getCoverageResult();
+      await outputFile.writeAsString(coverageResults, flush: true);
 
-    var parsedResult = await parseCoverage([outputFile], 1);
+      var parsedResult = await parseCoverage([outputFile], 1);
 
-    expect(parsedResult, contains(_sampleAppFileUri));
-    expect(parsedResult, contains(_isolateLibFileUri));
+      expect(parsedResult, contains(_sampleAppFileUri));
+      expect(parsedResult, contains(_isolateLibFileUri));
+    } finally {
+      await tempDir.delete(recursive: true);
+    }
   });
 }
 
@@ -93,19 +99,16 @@ Future<String> _getCoverageResult() async {
 }
 
 Future<String> _collectCoverage() async {
-  expect(await FileSystemEntity.isFile(_sampleAppPath), isTrue);
+  expect(await FileSystemEntity.isFile(testAppPath), isTrue);
 
-  // need to find an open port
-  var socket = await ServerSocket.bind(InternetAddress.ANY_IP_V4, 0);
-  int openPort = socket.port;
-  await socket.close();
+  var openPort = await getOpenPort();
 
   // run the sample app, with the right flags
   var sampleProcFuture = Process
       .run('dart', [
     '--enable-vm-service=$openPort',
     '--pause_isolates_on_exit',
-    _sampleAppPath
+    testAppPath
   ])
       .timeout(_timeout, onTimeout: () {
     throw 'We timed out waiting for the sample app to finish.';